+++ web/default/src/features/pricing/components/pricing-sidebar.test.tsx
@@ -16,77 +16,12 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 For commercial licensing, please contact https://github.com/MAX-API-Next/MAX-API/issues
 */
-import { act } from 'react'
-import { createRoot, type Root } from 'react-dom/client'
-import { createInstance } from 'i18next'
-import { JSDOM } from 'jsdom'
+import { createReactTestEnvironment } from '@/test/react'
 import assert from 'node:assert/strict'
 import { after, before, describe, test } from 'node:test'
-import { I18nextProvider } from 'react-i18next'
 import { PricingSidebar, type PricingSidebarProps } from './pricing-sidebar'
 
-const dom = new JSDOM('<!doctype html><html><body></body></html>', {
-  url: 'http://localhost/',
-})
-const i18n = createInstance()
-const globalKeys = [
-  'window',
-  'document',
-  'navigator',
-  'HTMLElement',
-  'Element',
-  'Node',
-  'Event',
-  'MouseEvent',
-  'KeyboardEvent',
-  'MutationObserver',
-  'getComputedStyle',
-  'requestAnimationFrame',
-  'cancelAnimationFrame',
-  'ResizeObserver',
-  'IS_REACT_ACT_ENVIRONMENT',
-] as const
-const previousDescriptors = new Map<
-  (typeof globalKeys)[number],
-  PropertyDescriptor | undefined
->()
-
-class ResizeObserverStub {
-  observe() {}
-  unobserve() {}
-  disconnect() {}
-}
-
-function setGlobal(key: (typeof globalKeys)[number], value: unknown) {
-  previousDescriptors.set(key, Object.getOwnPropertyDescriptor(globalThis, key))
-  Object.defineProperty(globalThis, key, {
-    configurable: true,
-    writable: true,
-    value,
-  })
-}
-
-function restoreGlobals() {
-  for (const key of [...globalKeys].reverse()) {
-    const descriptor = previousDescriptors.get(key)
-    if (descriptor) {
-      Object.defineProperty(globalThis, key, descriptor)
-    } else {
-      Reflect.deleteProperty(globalThis, key)
-    }
-  }
-}
-
-function createContainer() {
-  const container = dom.window.document.createElement('div')
-  dom.window.document.body.append(container)
-  return container
-}
-
-async function unmount(root: Root, container: HTMLElement) {
-  await act(async () => root.unmount())
-  container.remove()
-}
+const testEnv = createReactTestEnvironment()
 
 const sidebarProps: PricingSidebarProps = {
   quotaTypeFilter: 'all',
@@ -118,83 +53,29 @@ const sidebarProps: PricingSidebarProps = {
   onClearFilters: () => undefined,
 }
 
-before(async () => {
-  await i18n.init({
-    lng: 'en',
-    fallbackLng: 'en',
-    resources: { en: { translation: {} } },
-    interpolation: { escapeValue: false },
-  })
+before(() => testEnv.setup())
 
-  const window = dom.window
-  Object.defineProperty(window, 'matchMedia', {
-    configurable: true,
-    value: () => ({
-      matches: false,
-      media: '',
-      onchange: null,
-      addEventListener() {},
-      removeEventListener() {},
-      addListener() {},
-      removeListener() {},
-      dispatchEvent: () => false,
-    }),
-  })
-
-  setGlobal('window', window)
-  setGlobal('document', window.document)
-  setGlobal('navigator', window.navigator)
-  setGlobal('HTMLElement', window.HTMLElement)
-  setGlobal('Element', window.Element)
-  setGlobal('Node', window.Node)
-  setGlobal('Event', window.Event)
-  setGlobal('MouseEvent', window.MouseEvent)
-  setGlobal('KeyboardEvent', window.KeyboardEvent)
-  setGlobal('MutationObserver', window.MutationObserver)
-  setGlobal('getComputedStyle', window.getComputedStyle.bind(window))
-  setGlobal('requestAnimationFrame', (callback: FrameRequestCallback) =>
-    setTimeout(() => callback(Date.now()), 0)
-  )
-  setGlobal('cancelAnimationFrame', (handle: number) => clearTimeout(handle))
-  setGlobal('ResizeObserver', ResizeObserverStub)
-  setGlobal('IS_REACT_ACT_ENVIRONMENT', true)
-})
-
-after(() => {
-  restoreGlobals()
-  dom.window.close()
-})
+after(() => testEnv.teardown())
 
 describe('PricingSidebar auto route chains', () => {
   test('keeps configured auto route details collapsed until requested', async () => {
-    const container = createContainer()
-    const root = createRoot(container)
+    const view = await testEnv.render(<PricingSidebar {...sidebarProps} />)
 
     try {
-      await act(async () => {
-        root.render(
-          <I18nextProvider i18n={i18n}>
-            <PricingSidebar {...sidebarProps} />
-          </I18nextProvider>
-        )
-      })
-
-      const trigger = Array.from(container.querySelectorAll('button')).find(
-        (button) => button.textContent?.includes('Auto route chains')
-      )
+      const trigger = Array.from(
+        view.container.querySelectorAll('button')
+      ).find((button) => button.textContent?.includes('Auto route chains'))
 
       assert.ok(trigger)
       assert.equal(trigger.getAttribute('aria-expanded'), 'false')
-      assert.doesNotMatch(container.textContent || '', /Fast lane/)
+      assert.doesNotMatch(view.container.textContent || '', /Fast lane/)
 
-      await act(async () => {
-        trigger.click()
-      })
+      await view.click(trigger)
 
       assert.equal(trigger.getAttribute('aria-expanded'), 'true')
-      assert.match(container.textContent || '', /Fast lane/)
+      assert.match(view.container.textContent || '', /Fast lane/)
     } finally {
-      await unmount(root, container)
+      await view.unmount()
     }
   })
 })
